Create SearchInSortedRotatedArray
/*
The logic is to we can apply binary search, but here the elemnt are not completely sorted. So first we will find the point where from array in not sorted or the minimum value.

Step1.  Find the position where from array is rotated. Now from that point we can divide the array into two part for binary search.


 */